+++ src/main/java/myboot/web/PersonRestController.java
@@ -2,8 +2,11 @@ package myboot.web;
 
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,18 +20,23 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import myboot.Utility.IsEntityFromUserChecker;
 import myboot.dao.PersonRepository;
 import myboot.dto.PersonDTO;
 import myboot.model.Person;
 
 @RestController
 @RequestMapping("/api/persons")
+@Profile("usejwt")
 public class PersonRestController {
     @Autowired
     PersonRepository p_repo;
 
     @Autowired
     LocalValidatorFactoryBean validationFactory;
+
+    @Autowired
+    IsEntityFromUserChecker userChecker;
 
     private ModelMapper mapper = new ModelMapper();
 
@@ -51,7 +59,10 @@ public class PersonRestController {
     /// PostMapping
 
     @PostMapping("")
-    public PersonDTO postPerson(@RequestBody PersonDTO pdto) throws PersonNotFoundException {
+    public PersonDTO postPerson(@RequestBody PersonDTO pdto, HttpServletRequest req) throws PersonNotFoundException, IsEntityFromUserChecker.UserNotAllowedException {
+        if(! (userChecker.isUserAdmin(req) || userChecker.isMySelf(req, pdto.getSelf().getUsername())))
+            throw new IsEntityFromUserChecker.UserNotAllowedException();
+
         Person p = mapper.map(pdto, Person.class);
 
         Optional.ofNullable(p_repo.findById(p.getId()).isPresent() ? null : p).orElseThrow(PersonNotFoundException::new);
@@ -63,7 +74,10 @@ public class PersonRestController {
     /// PutMapping
 
     @PutMapping("")
-    public PersonDTO putPerson(@RequestBody PersonDTO pdto) throws PersonNotFoundException {
+    public PersonDTO putPerson(@RequestBody PersonDTO pdto, HttpServletRequest req) throws PersonNotFoundException, IsEntityFromUserChecker.UserNotAllowedException {
+        if(! (userChecker.isUserAdmin(req) || userChecker.isMyPerson(req, pdto.getId())))
+            throw new IsEntityFromUserChecker.UserNotAllowedException();
+        
         Person p = mapper.map(pdto, Person.class);
 
         Optional<Person> p_data = p_repo.findById(p.getId());
@@ -78,7 +92,10 @@ public class PersonRestController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deletePerson(@PathVariable int id) {
+    public void deletePerson(@PathVariable int id, HttpServletRequest req) throws IsEntityFromUserChecker.UserNotAllowedException {
+        if(! (userChecker.isUserAdmin(req) || userChecker.isMyPerson(req, id)))
+            throw new IsEntityFromUserChecker.UserNotAllowedException();
+
         p_repo.deleteById(id);
     }
 
@@ -86,7 +103,7 @@ public class PersonRestController {
     /// Exception
 
     @ResponseStatus(value = HttpStatus.NOT_FOUND)
-    public class PersonNotFoundException extends RuntimeException{
+    public static class PersonNotFoundException extends RuntimeException{
         private static final long serialVersionUID = 1L;
     }
 }
